chore : group and split viewcontrollers in mypage
